changed some db.open() statements to if statements for failsafing

+++ personrepository.cpp
@@ -19,12 +19,14 @@ PersonRepository::PersonRepository(string db_name, string conn_name)
 void PersonRepository::write(string line)
 {
     vector<string> fields = parse_delim_string(line, '|');
-    db.open();
-    QSqlQuery queryObj(db);
-    string queryString;
-    queryString = "INSERT INTO 'persons' (name, profession, description, birthyear, deathyear, sex) values('" + fields[0] + "', '" + fields[1] + "', '" + fields[2] + "', '" + fields[3] + "', '" + fields[4] + "', '" + fields[5] + "');";
-    QString qQueryString(queryString.c_str());
-    queryObj.exec(qQueryString);
+    if (db.open())
+    {
+        QSqlQuery queryObj(db);
+        string queryString;
+        queryString = "INSERT INTO 'persons' (name, profession, description, birthyear, deathyear, sex) values('" + fields[0] + "', '" + fields[1] + "', '" + fields[2] + "', '" + fields[3] + "', '" + fields[4] + "', '" + fields[5] + "');";
+        QString qQueryString(queryString.c_str());
+        queryObj.exec(qQueryString);
+    }
     db.close();
 }
 
@@ -33,7 +35,7 @@ vector<string> PersonRepository::read_entries()
 {
     vector<string> query_vect;
     string queryString;
-    if (this->db.open())
+    if (db.open())
     {
         QSqlQuery queryObj(db);
 
@@ -50,11 +52,13 @@ vector<string> PersonRepository::read_entries()
 void PersonRepository::remove(string column, string id)
 {
     string queryString;
-    db.open();
-    QSqlQuery queryObj(db);
-    queryString = "DELETE from persons WHERE " + column + "=" + id;
-    QString qQueryString(queryString.c_str());
-    queryObj.exec(qQueryString);
+    if (db.open())
+    {
+        QSqlQuery queryObj(db);
+        queryString = "DELETE from persons WHERE " + column + "=" + id;
+        QString qQueryString(queryString.c_str());
+        queryObj.exec(qQueryString);
+    }
     db.close();
 
 }
@@ -64,12 +68,14 @@ vector<string> PersonRepository::query_exact(string column, string data_query)
 {
     vector<string> query_vect;
     vector<string> result_vect;
-    db.open();
-    QSqlQuery queryObj(db);
-    string queryString = "SELECT * FROM persons WHERE " + column + " ='" + data_query + "'";
-    QString qQueryString(queryString.c_str());
-    queryObj.exec(qQueryString);
-    query_vect = from_db_to_vector("persons", queryObj);
+    if (db.open())
+    {
+        QSqlQuery queryObj(db);
+        string queryString = "SELECT * FROM persons WHERE " + column + " ='" + data_query + "'";
+        QString qQueryString(queryString.c_str());
+        queryObj.exec(qQueryString);
+        query_vect = from_db_to_vector("persons", queryObj);
+    }
     db.close();
 
     return query_vect;
@@ -80,12 +86,14 @@ vector<string> PersonRepository::query(string column, string data_query)
 {
     vector<string> query_vect;
     vector<string> result_vect;
-    db.open();
-    QSqlQuery queryObj(db);
-    string queryString = "SELECT * FROM persons WHERE " + column + " LIKE '%" + data_query + "%'";
-    QString qQueryString(queryString.c_str());
-    queryObj.exec(qQueryString);
-    query_vect = from_db_to_vector("persons", queryObj);
+    if (db.open())
+    {
+        QSqlQuery queryObj(db);
+        string queryString = "SELECT * FROM persons WHERE " + column + " LIKE '%" + data_query + "%'";
+        QString qQueryString(queryString.c_str());
+        queryObj.exec(qQueryString);
+        query_vect = from_db_to_vector("persons", queryObj);
+    }
     db.close();
 
     return query_vect;
